fix: some clean up around buttons
