+++ client/src/components/TrackGroup/PurchaseAlbumModal.tsx
@@ -2,7 +2,6 @@ import { css } from "@emotion/css";
 import { useQuery } from "@tanstack/react-query";
 import { ArtistButton } from "components/Artist/ArtistButtons";
 import CurrencyCoinIcon from "components/common/CurrencyCoinIcon";
-import { FixedButton } from "components/common/FixedButton";
 import Modal from "components/common/Modal";
 import BuyTrackGroup from "components/TrackGroup/BuyTrackGroup";
 import { queryArtist } from "queries";
@@ -65,13 +64,14 @@ const PurchaseAlbumModal: React.FC<{
     : "buyingTrackGroup";
 
   const button = fixed ? (
-    <FixedButton
+    <ArtistButton
       onClick={() => setIsPurchasingAlbum(true)}
       rounded
+      variant="pill"
       size="compact"
     >
       {t(preOrderOrBuyText)}
-    </FixedButton>
+    </ArtistButton>
   ) : (
     <ArtistButton
       type="button"
